+++ src/simmate/website/htmx/components/mixins/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 
 from .molecule import MoleculeInput
+from .structure import StructureInput
 from .user import UserInput